Provide an API option to avoid exiting
When the driver needs to exit, e.g. due to an error, or as the result of
an internals command, it will just call sys.exit.

When testing, or with potencial advanced uses of a custom driver, this
might be undesirable. Setting the 'exit' keyword argument of
subdue.sub.main to False will prevent the driver from exiting. If an
exit condition occurs, then the main function will return a SystemExit
exception object.