Stop using sub's name in environment vars
All subdue provided environment variables now share the `SUB_` prefix.
Originally the sub name was there so that nested sub invocations (yes,
that is possible, a command could be a sub on itself) would not confuse
the variables. However, by sharing the name, and making sure all of them
are set when running the driver, we can guarantee that the values of the
variables seen by the command have been set by the closes calling sub
driver in the call chain.